Testing ActiveJob with different argument types
I was getting quite confused with how to use assert_enqueued_with in
my tests where they were exercising a job that used a hash for
arguments. I was sure there was a bug in ActiveJob so I came to write
some tests to find it. In doing so I learned how it was supposed to
work. Hopefully these tests might help someone else understand if
they come looking.

* consolidate job serialization tests to one file

* make explicit expectations of jobs with hash args in tests

* make explicit examples of how to use assert_enqueued_with with jobs
  that take keyword or hash args